requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
rfc3986==1.5.0
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
